Admin Task View Page
Page for users that are admins of a team to view tasks that users have marked as completed. Currently clicking a task takes you to the completed task detail page which gives a button to mark the tasks as not complete. Another button is needed to mark the task as valid so that it is removed from all other task lists.